added security rule to makefile, switched sclean and fclean rule, added permissionchangeing to all rule for linux, added modules in nestjs 4 installation dcom, added swagger mysql and phpmyadmin 4 api handling and the last 2 for database development if needed, added command for empty database folder in makefile, added Notice with all addresses

+++ Makefile
@@ -5,6 +5,8 @@ SRC		:=	$(SRC_D)$(SRC_F)
 
 ENV		:=	--env-file $(SRC_D).env
 DB_D	:=	./data/sql
+OS		:=	$(shell uname)
+
 
 ###			###			COLORS			###			###
 RED		=	\033[1;31m
@@ -13,10 +15,21 @@ YELL	=	\033[1;33m
 BLUE	=	\033[1;34m
 WHITE	=	\033[0m
 
+###			###			SECURITY		###			###
+ifndef CONFIRM
+  CONFIRM_MESSAGE := "$(WHITE)Are you sure you want to continue? This will $(RED)DELETE ALL configs and databases!$(WHITE) (Write 'CONFIRM=1 make sclean' to continue.)"
+endif
+
 ###			###			RULES			###			###
 #Build changes or all if nothing is builded and run
 all:
-	@mkdir -p $(DB_D)
+	@mkdir -p data $(DB_D) data/myadmin data/mysql data/pgadmin
+	@mkdir -p $(DB_D)/pg_notify $(DB_D)/pg_tblspc $(DB_D)/pg_replslot $(DB_D)/pg_twophase $(DB_D)/pg_snapshots $(DB_D)/pg_logical/snapshots $(DB_D)/pg_logical/mappings $(DB_D)/pg_commit_ts
+ifeq ($(OS), Darwin)
+	bash -c "chmod 600 data/pgadmin/pgadmin4.db && chown -R ${USER}:2021_heilbronn data/pgadmin"
+else
+	bash -c "chmod 600 data/pgadmin/pgadmin4.db && sudo chown -R 5050:5050 data/pgadmin"
+endif
 	docker-compose -f $(SRC) $(ENV) up
 	@echo "$(GREEN)Build changes and/or new containers.$(WHITE)"
 
@@ -52,12 +65,17 @@ clean: stop
 	@echo "$(BLUE)Stopped all containers and clean images.$(WHITE)"
 
 fclean: clean
-	-@sudo rm -rf $(DB_D) && mkdir -p $(DB_D); true;
-	@echo "$(BLUE)Cleaned sql and wordpress data.$(WHITE)"
-
-sclean: fclean
 	-@docker system prune -a --volumes --force
 	@echo "$(BLUE)Hardcleaned docker.$(WHITE)"
+
+sclean: fclean
+ifdef CONFIRM
+	@echo "$(YELL)Continuing with the operation...$(WHITE)"
+	-@sudo rm -rf $(DB_D) && mkdir -p $(DB_D); true;
+	@echo "$(BLUE)Cleaned all data.$(WHITE)"
+else
+	@echo $(CONFIRM_MESSAGE)
+endif
 
 #stop all containers, force rebuild and start it
 re: stop build all
